21. Fix a trouble with FilterRegistrationBean overriding
